+++ ChangeLog.md
@@ -5,7 +5,8 @@ XP Compiler ChangeLog
 
 ## 4.3.0 / ????-??-??
 
-* Fixed emitting `switch` statements - @thekid
+* Fixed emitting `switch` statements and case labels' ambiguity w/ goto
+  (@thekid)
 * Fixed an operator precedence problem causing incorrect nesting in the
   parsed AST for unary prefix operators.
   (@thekid)